requirements.txt
numpy
scipy
pytorch
zounds
librosa
lmdb
conjure